requirements.txt
streamlit
google-generativeai
feedparser
newspaper3k